requirements.txt
Faker==30.8.2
python-dotenv==1.2.1
psycopg2>=2.9.11
ruff==0.8.2
